inherit classes for tests to not repeat code
